학습: Re축 이미지 float16 저장으로 13년치 데이터 OOM 수정
13년치 ASOS(약 130만 표본)로 재학습 시 Re축 (4,32,32) 이미지를
float32로 전량 미리 만들어 두면 약 22GB, 거기에 리스트+np.stack
중간본이 겹쳐 WSL2/Docker 메모리 한도(31GB)를 넘겨 OOM-kill됐다
(2026-08-15 실측, exit 137). InterpolatedFieldCollector.get_batch가
float16 배열을 바로 채워 넣게 바꾸고, 모델에 넣기 직전(train.py 및
X_img를 재사용하는 진단 스크립트들)에서 float32로 캐스팅한다.

+++ backtest_accuracy.py
@@ -58,7 +58,7 @@ def main():
             sl = slice(b, b + BATCH)
             pred = model(
                 num_x=ds.X_num[sl].to(DEVICE),
-                img_x=ds.X_img[sl].to(DEVICE) if ds.X_img is not None else None,
+                img_x=ds.X_img[sl].to(DEVICE).float() if ds.X_img is not None else None,
                 txt_x=ds.X_txt[sl].to(DEVICE) if ds.X_txt is not None else None,
             )
             preds.append(pred.cpu().numpy())
